backend/requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# Dependencias del backend — Sistema de usuarios sintéticos Moeve
# ─────────────────────────────────────────────────────────────────────────────

# Framework web y servidor ASGI
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Validación de datos y schemas
pydantic==2.9.2
pydantic-settings==2.5.2

# Variables de entorno
python-dotenv==1.0.1

# Base de datos SQLite asíncrona
aiosqlite==0.20.0
databases==0.9.0

# Proveedores LLM
openai==1.50.0
anthropic==0.34.2
httpx==0.27.2          # Cliente HTTP para Ollama y timeouts

# Generación de PDFs desde HTML
weasyprint==62.3
pydyf==0.11.0          # Pinned: 0.12.x rompe weasyprint 62.3 (AttributeError en stream.transform)
jinja2==3.1.4          # Motor de plantillas para el HTML del PDF

# Rate limiting
slowapi==0.1.9

# Subida de archivos multipart
python-multipart==0.0.9

# Parseo de archivos de perfiles y briefs
PyYAML==6.0.2

# Formateo de código (solo desarrollo)
black==24.8.0
isort==5.13.2
